Move `toolbar` => `Toolbar`
