+++ src/annotator/components/Toolbar.js
